requirements.txt
et-xmlfile==1.1.0
numpy==1.23.1
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
